+++ src/geometry2D/Point.java
@@ -7,19 +7,19 @@ package geometry2D;
  */
 public class Point {
 
-    final private float x;
-    final private float y;
+    final private double x;
+    final private double y;
 
-    public Point(float x, float y){
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public float getX(){
+    public double getX(){
         return x;
     }
 
-    public float getY(){
+    public double getY(){
         return y;
     }
 }
